% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_adaptive_hlasso.R
\name{fit.adaptive.hlasso}
\alias{fit.adaptive.hlasso}
\title{Fit adaptive hierarchical lasso}
\usage{
fit.adaptive.hlasso(x.t, y, beta.ini, m, n, p, info.half, BIC = FALSE,
  max.iter = 50, print = FALSE)
}
\arguments{
\item{x.t}{Design matrix or pseudo-design matrix where each column has been
multiplied by abs(beta) corresponding to it}

\item{y}{Outcome or pseudo-outcome matrix}

\item{m}{Number of outcomes}

\item{n}{Number of observations}

\item{p}{Number of predictors}

\item{info.half}{Half Fisher information matrix for beta}

\item{BIC}{Logical, if FALSE, modified BIC criterion is used}

\item{max.iter}{Maximum number of iterations in iterative procedure}

\item{print}{Logical indicating whether to print intermediate updates}
}
\value{
A list with the following elements:\itemize{
  \item \code{lasso.beta}: the non-hierarchical adaptive lasso estimate
  \item \code{hlasso.beta}: the adaptive hierarchical lasso estimate
  \item \code{num.iter}: the number of iterations used
  \item \code{d}: the estimate of d
  \item \code{alpha}: the estimate of alpha
}
}
\description{
Fit adaptive hierarchical lasso
}
\details{
Fits an adaptive version of the hierarchical lasso
suitable for multivariate mixed outcomes. Variable selection is performed
to both remove uninformative predictors and remove unimportant predictor-outcome
relationships. Tuning parameter selection is done via a BIC-style criterion.
}
\seealso{
\code{\link[lars]{lars}}
}

