% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_ini.R
\name{fit.ini}
\alias{fit.ini}
\title{Produce pseudo x and y based on marginal proportion odds model fits and quadratic}
\usage{
fit.ini(x, y, xname, yname = "y", p)
}
\arguments{
\item{x}{Named matrix of predictors}

\item{y}{n x m matrix of outcomes}

\item{xname}{String to identify predictors in x}

\item{yname}{String to identify outcomes in y}

\item{p}{Number of predictors (doesn't need to be same as number of columns of x)



#'}
}
\value{
A list with the following elements:\itemize{
  \item \code{pseudoY}: pseudo-outcome for quadratic expansion of the likelihood
  \item \code{betahat}: coefficients corresponding to the NPMLE of finite-dimensional parameters
  \item \code{alphahat}: coefficients corresponding to the NPMLE of infinite dimensional parameters
  \item \code{pseudoX}: pseudo-predictors for quadratic expansion
  \item \code{pseudoX.t}: adaptive version of pseudo-predictors
  \item \code{inv.info}: inverse Fisher information from  marginal models
  \item \code{inv.info.beta}: just the part of the inverse information corresponding to beta
  \item \code{beta.inds}: indices in the coefficient vectors corresponding to beta
  \item \code{alpha.inds}: indices in the coefficient vectors corresponding to the infinite-dimensional parameter
  \item \code{b0.inds}: indices in the coefficient vectors corresponding to nuisance covariates
  \item \code{coefs}: full coefficient vector
}
}
\description{
Produce pseudo x and y based on marginal proportion odds model fits and quadratic
}

