% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_fns.R
\name{generate.x}
\alias{generate.x}
\title{Generate x values in \{0,1,2\}.}
\usage{
generate.x(n, p.low, p.high, p.x)
}
\arguments{
\item{n}{Sample size.}

\item{p.low}{Lower bound for minor allele frequency.}

\item{p.high}{Upper bound for minor allele frequency.}

\item{p.x}{Number of predictors.}
}
\value{
x matrix
}
\description{
Generate x values in \{0,1,2\}.
}

