% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/perturb_beta.R
\name{perturb.beta}
\alias{perturb.beta}
\title{Get perturbed version of the nonparametric maximum likelihood estimator}
\usage{
perturb.beta(num.ptb, x, y, psd, influence.fn = po.influence)
}
\arguments{
\item{num.ptb}{Number of perturbations to perform}

\item{x}{n x p original design matrix}

\item{y}{List of length m containing outcomes}

\item{psd}{Pseudo-data generated from x and y based on marginal models}

\item{influence.fn}{Function to obtain the influence function of beta based on psd}
}
\value{
n x num.ptb matrix with perturbed estimates
}
\description{
Get perturbed version of the nonparametric maximum likelihood estimator
}
\details{
Perturbed estimates are computed based on a beta distribution chosen to have
first three moments equal to 1 (and then mean-shifted).
}

