% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smrt_stepdown.R
\name{smrt.stepdown}
\alias{smrt.stepdown}
\title{Perform stepdown testing on parameter estimates for each predictor}
\usage{
smrt.stepdown(beta, betastar, inv.info = NULL, inv.info.beta = NULL, z.th,
  p, m)
}
\arguments{
\item{beta}{Vector of estimates}

\item{betastar}{Matrix of perturbed estimates}

\item{inv.info}{Full Fisher information matrix}

\item{inv.info.beta}{Sub-matrix of Fisher information corresponding to \code{beta}.}

\item{z.th}{Threshold for rejection, either scalar or p-length vector.}

\item{get.u}{Function to obtain the influence function of beta based on \code{psd}.}
}
\value{
Indices of beta corresponding to rejected tests.
}
\description{
Perform stepdown testing on parameter estimates for each predictor
}
\details{
Estimates should be arranged such that all estimates for first outcome
come first, then all estimates for second outcomes, and so on. Only one of
inv.info and inv.info.beta need to be supplied.
}

