% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smrt_stepdown.R
\name{stepdown}
\alias{stepdown}
\title{Basic stepdown function}
\usage{
stepdown(z, zstar, thr)
}
\arguments{
\item{z}{Index vector of estimates}

\item{zstar}{Matrix of perturbed estimates}

\item{thr}{Threshold for rejection}
}
\value{
Indices corresponding to rejected tests.
}
\description{
Basic stepdown function
}

